Refactor Intervention plan directory structure
