+++ src/containers/pages/InterventionPlan/IRSPlan/serviceCalls.ts
@@ -1,8 +1,8 @@
-import { OPENSRP_PLANS } from '../../../../../constants';
-import { displayError } from '../../../../../helpers/errors';
-import { extractPlanRecordResponseFromPlanPayload } from '../../../../../helpers/utils';
-import { OpenSRPService } from '../../../../../services/opensrp';
-import { fetchPlanRecords, PlanPayload } from '../../../../../store/ducks/plans';
+import { OPENSRP_PLANS } from '../../../../constants';
+import { displayError } from '../../../../helpers/errors';
+import { extractPlanRecordResponseFromPlanPayload } from '../../../../helpers/utils';
+import { OpenSRPService } from '../../../../services/opensrp';
+import { fetchPlanRecords, PlanPayload } from '../../../../store/ducks/plans';
 
 type UUID = string;
