+++ src/containers/pages/InterventionPlan/PlanningView/tests/fixtures.ts
